fix userName in csr is incorrect if cluster name is long
