+++ pkg/bootstrap/bootstrapkubeconfig_test.go
@@ -1081,34 +1081,6 @@ func TestGetValidCertificatesFromURL(t *testing.T) {
 	}
 }
 
-func TestGetBootstrapSAName(t *testing.T) {
-	cases := []struct {
-		name           string
-		clusterName    string
-		expectedSAName string
-		managedCluster *clusterv1.ManagedCluster
-	}{
-		{
-			name:           "short name",
-			clusterName:    "123456789",
-			expectedSAName: "123456789-bootstrap-sa",
-		},
-		{
-			name:           "long name",
-			clusterName:    "123456789-123456789-123456789-123456789-123456789-123456789",
-			expectedSAName: "123456789-123456789-123456789-123456789-123456789--bootstrap-sa",
-		},
-	}
-
-	for _, c := range cases {
-		t.Run(c.name, func(t *testing.T) {
-			if c.expectedSAName != GetBootstrapSAName(c.clusterName) {
-				t.Errorf("expected sa %v, but got %v", c.expectedSAName, GetBootstrapSAName(c.clusterName))
-			}
-		})
-	}
-}
-
 func TestGetProxySettings(t *testing.T) {
 	tests := []struct {
 		name             string
