+++ pkg/helpers/csr_test.go
@@ -6,6 +6,7 @@ import (
 	"github.com/stolostron/managedcluster-import-controller/pkg/constants"
 	certificatesv1 "k8s.io/api/certificates/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	clusterv1 "open-cluster-management.io/api/cluster/v1"
 )
 
 func Test_getClusterName(t *testing.T) {
@@ -69,6 +70,34 @@ func Test_getClusterName(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			if gotClusterName := GetClusterName(tt.args.csr); gotClusterName != tt.wantClusterName {
 				t.Errorf("getClusterName() = %v, want %v", gotClusterName, tt.wantClusterName)
+			}
+		})
+	}
+}
+
+func TestGetBootstrapSAName(t *testing.T) {
+	cases := []struct {
+		name           string
+		clusterName    string
+		expectedSAName string
+		managedCluster *clusterv1.ManagedCluster
+	}{
+		{
+			name:           "short name",
+			clusterName:    "123456789",
+			expectedSAName: "123456789-bootstrap-sa",
+		},
+		{
+			name:           "long name",
+			clusterName:    "123456789-123456789-123456789-123456789-123456789-123456789",
+			expectedSAName: "123456789-123456789-123456789-123456789-123456789--bootstrap-sa",
+		},
+	}
+
+	for _, c := range cases {
+		t.Run(c.name, func(t *testing.T) {
+			if c.expectedSAName != GetBootstrapSAName(c.clusterName) {
+				t.Errorf("expected sa %v, but got %v", c.expectedSAName, GetBootstrapSAName(c.clusterName))
 			}
 		})
 	}
